requirements.txt
requests>=2.28.2
pytest>=7.2.2
python-dotenv>=1.0.0